% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{nz_connect}
\alias{nz_connect}
\title{nz_connect}
\usage{
nz_connect(dsn = "100.0.0.1:666", uid = "", pwd = "", database = "",
  classPath = "C://JDBC//DIRECTORY", dType = "jdbc")
}
\arguments{
\item{dsn}{}

\item{uid}{}

\item{pwd}{}

\item{database}{}

\item{JDBC}{driver path (nzjdbc.jar path)}

\item{dtype}{type of Driver (JDBC is recommended)}
}
\description{
Connects to netezza using RODBC or RJDBC.
}

