% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{nz_export}
\alias{nz_export}
\title{nz_export}
\usage{
nz_export(query, fetchSize = 100000, verbose = TRUE, file = "export.txt",
  col_names = TRUE, ...)
}
\arguments{
\item{q}{query to be exported}

\item{fetchSize}{}

\item{file}{}

\item{requires}{a nzcon Java Connection}
}
\description{
Export the result of a query to a .txt file in chunks
}

