% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{nz_sendquery}
\alias{nz_sendquery}
\title{nz_sendquery}
\usage{
nz_sendquery(query = NULL)
}
\arguments{
\item{q}{query to be exported}

\item{fetchSize}{}

\item{file}{}

\item{requires}{a nzcon Java Connection}
}
\description{
Sends a query to the database and but doesn't get the result.
Solves the "white space issues" RJDBC has by adding a trim() call to each field.
}

