% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{td2nz}
\alias{td2nz}
\title{Exports a table from teradata into Netezza}
\usage{
td2nz(td_query = NULL, new_nz_table = "td2nz")
}
\arguments{
\item{td_query}{Teradata Query}

\item{new_nz_table}{DATABASE.USER.TABLENAME}

\item{data_type}{All collums datatype, defaults to VARCHAR(255)}

\item{requires}{a netezza and teradata JDBC connection}
}
\description{
Exports a table from teradata into Netezza
}

