% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{td_export}
\alias{td_export}
\title{Export tables from Teradata to a .txt file}
\usage{
td_export(q, fetchSize = 100000, verbose = T, file = "export.txt",
  col_names = TRUE, ...)
}
\arguments{
\item{q}{query to be exported}

\item{fetchSize}{Size of chunks to be received in memory}

\item{file}{.txt file to be exported locally}
}
\description{
Requires a tdConnection object in the Gloval Environment
}

