% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{td_fun_export}
\alias{td_fun_export}
\title{Applies a function while exporting table from Teradata}
\usage{
td_fun_export(q, fetchSize = 100000, verbose = T, FUN = NULL,
  file = "export.txt", col_names = TRUE, ...)
}
\arguments{
\item{q}{query to be exported}

\item{fetchSize}{Size of chunks to be received in memory}

\item{file}{.txt file to be exported locally}
}
\description{
Requires a tdConnection object in the Gloval Environment
}

